detect changes in available authentication
